refactor(react-native): android app start timeout should be a `long`

+++ packages/platforms/react-native/android/src/main/java/com/bugsnag/android/performance/ReactNativeAppStartPlugin.java
@@ -17,7 +17,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 public class ReactNativeAppStartPlugin implements Plugin {
 
-  private static final int DEFAULT_SPAN_BLOCK_TIMEOUT_MS = 5000;
+  private static final long DEFAULT_SPAN_BLOCK_TIMEOUT_MS = 5000;
 
   // Container class to ensure atomic updates of related fields
   private static class ViewLoadCondition {
@@ -34,13 +34,13 @@ public class ReactNativeAppStartPlugin implements Plugin {
 
   private final AtomicReference<ViewLoadCondition> viewLoadCondition = new AtomicReference<>(null);
   private volatile boolean appStartComplete = false;
-  private final int spanBlockTimeoutMs;
+  private final long spanBlockTimeoutMs;
 
   public ReactNativeAppStartPlugin() {
     this(DEFAULT_SPAN_BLOCK_TIMEOUT_MS);
   }
 
-  public ReactNativeAppStartPlugin(int timeoutMs) {
+  public ReactNativeAppStartPlugin(long timeoutMs) {
     this.spanBlockTimeoutMs = timeoutMs;
   }
